Add files via upload
'Fake news detection' project was done in which one has to identify which news is fake and which is not by using some NLP based concepts. This project has utilized libraries like IDFTF Vectorizer to identify which words are occurring in the news most of the time. The project has also used Passive Aggressive algorithm.